backend/requirements.txt
torch
torchvision
Pillow
numpy
openai-clip @ git+https://github.com/openai/CLIP.git
fastapi
uvicorn
python-multipart
